movies controller code cleanup, added a new Utils module
